+++ Uteis/Attributes/AttributeCorStatus.cs
