--- Uteis/ResponseFromAspNet.cs
@@ -1,7 +1,0 @@
-﻿namespace APICeomedAplicacoes.Uteis
-{
-    public class ResponseFromAspNet
-    {
-        public List<object> errors { get; set; }
-    }
-}
